desktop-client/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn==0.32.0
psutil==6.0.0
pynput==1.7.7
sqlalchemy==2.0.32
aiosqlite==0.20.0
pydantic==2.9.2
python-dotenv==1.0.1

# Development dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.8.0
flake8==7.1.1
mypy==1.11.2

# Optional ML dependencies
# scikit-learn==1.5.2
# pandas==2.2.3
# numpy==2.1.1